cloud_requirements.txt
absl-py==0.1.9
astor==0.6.2
bleach==1.5.0
cycler==0.10.0
decorator==4.2.1
enum34==1.1.6
gast==0.2.0
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.3
Markdown==2.6.11
matplotlib==2.1.2
networkx==2.0
numpy==1.14.0
opencv-python==3.4.0.12
Pillow==5.0.0
pkg-resources==0.0.0
progressbar2==3.34.3
protobuf==3.5.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-utils==2.2.0
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
scikit-image==0.13.1
scipy==1.0.0
six==1.11.0
tensorflow==1.5.0rc1
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
http://download.pytorch.org/whl/cu90/torch-0.3.0.post4-cp36-cp36m-linux_x86_64.whl 
torchvision==0.2.0
tqdm==4.19.5
Werkzeug==0.14.1
